+++ pax-properties-coproduct/src/lib.rs
@@ -24,7 +24,7 @@ pub enum PropertiesCoproduct {
     RepeatList(Vec<Rc<RefCell<PropertiesCoproduct>>>),
     RepeatItem(Rc<PropertiesCoproduct>, usize),
     //generated
-    i64(i64), //used by range for repeat (0..10)
+    isize(isize), //used by range for repeat (0..10)
 
     Rectangle(pax_example::pax_types::pax_std::primitives::RectangleProperties),
     Group(pax_example::pax_types::pax_std::primitives::GroupProperties),
@@ -36,7 +36,7 @@ pub enum PropertiesCoproduct {
 pub enum TypesCoproduct {
     //core (?)
     f64(f64),
-    i64(i64),
+    isize(isize),
     usize(usize), //used by Placeholder for index
     Vec_Rc_PropertiesCoproduct___(Vec<Rc<PropertiesCoproduct>>),
     DeeperStruct(pax_example::DeeperStruct),
